requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cssselect==1.1.0
fake-useragent==0.1.11
idna==3.3
importlib-metadata==4.11.3
lxml==4.8.0
mysql-connector-python==8.0.29
parse==1.19.0
protobuf==3.20.1
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
pyTelegramBotAPI==4.5.0
requests==2.27.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.0
urllib3==1.26.9
w3lib==1.22.0
websockets==10.3
zipp==3.8.0
